Move fancy AWT stroke to com.jhlabs.awt package.
Add few comments.
